refactor: rename SM120 grouped GEMM csrc to SM12x
Keep grouped_mm Python packages on the SM120 legacy import paths while moving the C++/JIT backend root to cute_sm12x_gemm.

Verification: pre-commit run -a; pytest -q tests/grouped_mm/test_cute_sm120_mxfp8.py tests/grouped_mm/test_cute_sm120_fp8.py.
