+++ flashinfer/jit/cute_sm120_mxfp8_groupwise.py
@@ -30,19 +30,19 @@ def gen_gemm_sm120_module_cute_mxfp8() -> JitSpec:
 
     Bundles the cute SM120 MXFP8 and FP8 float-scale groupwise runners with their
     public MoE entries (``moe_gemm_mxfp8_nt_groupwise`` / ``moe_gemm_fp8_nt_groupwise``)
-    in a single ``.so``. All source is in-tree under ``csrc/cute_sm120_mxfp8_groupwise/
-    {cute_sm120_{mxfp8,fp8}_runner.{h,cu}, cute_sm120_{mxfp8,fp8}_op{,_jit_binding}.cu,
+    in a single ``.so``. All source is in-tree under ``csrc/cute_sm12x_gemm/
+    {cute_sm12x_{mxfp8,fp8}_runner.{h,cu}, cute_sm12x_{mxfp8,fp8}_op{,_jit_binding}.cu,
     sm120_blockscaled/, sm120_blockscaling/, sm120_common/}``;
     the kernel uses flashinfer's own ``3rdparty/cutlass``.
     """
-    csrc_dir = jit_env.FLASHINFER_CSRC_DIR / "cute_sm120_mxfp8_groupwise"
+    csrc_dir = jit_env.FLASHINFER_CSRC_DIR / "cute_sm12x_gemm"
     source_paths = [
-        csrc_dir / "cute_sm120_mxfp8_runner.cu",
-        csrc_dir / "cute_sm120_mxfp8_op.cu",
-        csrc_dir / "cute_sm120_mxfp8_op_jit_binding.cu",
-        csrc_dir / "cute_sm120_fp8_runner.cu",
-        csrc_dir / "cute_sm120_fp8_op.cu",
-        csrc_dir / "cute_sm120_fp8_op_jit_binding.cu",
+        csrc_dir / "cute_sm12x_mxfp8_runner.cu",
+        csrc_dir / "cute_sm12x_mxfp8_op.cu",
+        csrc_dir / "cute_sm12x_mxfp8_op_jit_binding.cu",
+        csrc_dir / "cute_sm12x_fp8_runner.cu",
+        csrc_dir / "cute_sm12x_fp8_op.cu",
+        csrc_dir / "cute_sm12x_fp8_op_jit_binding.cu",
     ]
 
     nvcc_flags = current_compilation_context.get_nvcc_flags_list(
@@ -50,7 +50,7 @@ def gen_gemm_sm120_module_cute_mxfp8() -> JitSpec:
     )
 
     return gen_jit_spec(
-        "cute_sm120_mxfp8_groupwise",
+        "cute_sm12x_gemm",
         source_paths,
         extra_cuda_cflags=[*nvcc_flags, "-DCUTLASS_ENABLE_GDC_FOR_SM100=1"],
         extra_include_paths=[jit_env.FLASHINFER_CSRC_DIR],
